release!: 1.2.2 (#6)
* Refactoring + LSP

* Feat/update templates (#3)

* Update feature_request.md

* Update bug_report.md

* Update pull_request.md

* Rename .github/ISSUE_TEMPLATE/pull_request.md to .github/pull_request_template.md

* Update issue templates

* Create CODE_OF_CONDUCT.md

* Create CONTRIBUTING.md

* feat!: sqlite (#5)

* SQLite Implementation

!! SQLite does not provide Stored Procedures. Need more work on SQLiteCo object.

* fix!: Namespaces & connectionString checks

- IConnection and DbType renamed,
- Connection removed.

BREAKING CHANGE: new approad of connection instantiation. Doc coming...

* Create build-and-test.pullrequest.yml (#7)

* Create build-and-test.pullrequest.yml

* Update build-and-test.pullrequest.yml
