+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,55 +1,38 @@
 ---
 name: Bug report
-about: Créez un rapport pour nous aider à nous améliorer
-title: ''
+about: Create a report to help us improve
+title: "[BUG]"
 labels: bug
 assignees: ''
 
 ---
 
-### Preflight Checklist
+**Describe the bug**
+A clear and concise description of what the bug is.
 
-<!-- Please ensure you've completed the following steps by replacing [ ] with [x]-->
+**To Reproduce**
+Steps to reproduce the behavior:
+1. Go to '...'
+2. Click on '....'
+3. Scroll down to '....'
+4. See error
 
-- [ ] J'ai recherché dans le suivi des problèmes un problème qui correspond à celui que je souhaite déposer, sans succès.
+**Expected behavior**
+A clear and concise description of what you expected to happen.
 
-#
+**Screenshots**
+If applicable, add screenshots to help explain your problem.
 
-### Décrivez le bogue
+**Desktop (please complete the following information):**
+ - OS: [e.g. iOS]
+ - Browser [e.g. chrome, safari]
+ - Version [e.g. 22]
 
-Une description claire et concise de ce qu'est le bogue.
+**Smartphone (please complete the following information):**
+ - Device: [e.g. iPhone6]
+ - OS: [e.g. iOS8.1]
+ - Browser [e.g. stock browser, safari]
+ - Version [e.g. 22]
 
-#
-
-### Reproduire
-
-Étapes pour reproduire le comportement :
-
-1. Accédez à «...»
-2. Cliquez sur "...."
-3. Faites défiler jusqu'à «....»
-4. Voir erreur
-
-#
-
-### Qu'est-ce qui était attendu?
-
-Une description claire et concise de ce à quoi vous vous attendiez.
-
-#
-
-### Captures d'écran
-
-<!-- While providing screenshots, delete the text below !-->
-
-<!-- try as much as possible to explain each change in each screenshot !-->
-
-<!-- Don't forget to delete the p element while providing screeshots !-->
-
-<p align="center">Le cas échéant, ajoutez des captures d'écran pour expliquer votre problème.</p>
-
-#
-
-### Contexte supplémentaire
-
-Ajoutez tout autre contexte sur le problème ici (incluez les numéros de commit et les branches le cas échéant)
+**Additional context**
+Add any other context about the problem here.
