--- .github/ISSUE_TEMPLATE/pull_request.md
@@ -1,39 +1,0 @@
-
-## Description
-
-Veuillez inclure un résumé de la modification et du problème résolu. Veuillez également inclure la motivation et le contexte pertinents. Répertoriez toutes les dépendances requises pour cette modification.
-
-Fixes # (issue)
-
-## Type de changement
-
-Veuillez supprimer les options qui ne sont pas pertinentes.
-
-- [ ] Bug fix (non-breaking change which fixes an issue)
-- [ ] New feature (non-breaking change which adds functionality)
-- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
-- [ ] This change requires a documentation update
-
-# Comment cela a-t-il été testé ?
-
-Veuillez décrire les tests que vous avez exécutés pour vérifier vos modifications. Fournissez des instructions afin que nous puissions reproduire. Veuillez également énumérer tous les détails pertinents pour votre configuration de test
-
-- [ ] Test A
-- [ ] Test B
-
-**Configuration du test**:
-* Firmware version:
-* Hardware:
-* Toolchain:
-* SDK:
-
-# Liste de contrôle :
-
-- [ ] Mon code suit les directives de style de ce projet
-- [ ] J'ai effectué une auto-révision de mon propre code
-- [ ] J'ai commenté mon code, en particulier dans les zones difficiles à comprendre
-- [ ] J'ai apporté les modifications correspondantes à la documentation
-- [ ] Mes modifications ne génèrent aucun nouvel avertissement
-- [ ] J'ai ajouté des tests qui prouvent que mon correctif est efficace ou que ma fonctionnalité fonctionne
-- [ ] Les tests unitaires nouveaux et existants passent localement avec mes modifications
-- [ ] Toutes les modifications dépendantes ont été fusionnées et publiées dans les modules en aval
